Conjunction.toCNF() raises AssertionError
There was a problem in Conjunction.toCNF() method. If all clauses are True (e.g. `(!smoke(A) => smoke(A)) ^ (!smoke(A) => smoke(A))`), then, at the end of method, creating object of Conjunction will give error as it require at least 2 clauses. So, instead of creating object of Conjunction, we can create object of TrueFalse with value True.